fix: 過剰ファイル分割の抑制
